# src/app/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Simple Auth",
    "description": "Simple auth api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:3000/v1"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["user"],
        "summary": "register user",
        "description": "register new user",
        "operationId": "RegisterUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/auth/checkpassword": {
      "post": {
        "tags": ["user"],
        "summary": "check user password",
        "description": "register new user",
        "operationId": "CheckPassword",
        "security": [{ "basicAuth": [] }],
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    }
  },
  "components": {
    "schemas": {
      "CommonResponse": {
        "title": "CommonResponse",
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "fail", "error"],
            "description": "Response Status",
            "example": "error"
          },
          "code": {
            "type": "integer",
            "format": "int64",
            "example": 500,
            "description": "Response Code"
          },
          "message": {
            "maxLength": 255,
            "type": "string",
            "description": "Message Text.",
            "example": "Missing field value"
          }
        }
      },
      "DataResponse": {
        "title": "DataResponse",
        "allOf": [
          { "$ref": "#/components/schemas/CommonResponse" },
          { "type": "object", "properties": { "data": { "type": "string" } } }
        ]
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}